pet-store-be/requirements/requirements.txt
fastapi
uvicorn
pydantic[email]
svix-ksuid==0.6.0
python-multipart==0.0.5
PyJWT
pymysql
cryptography
sqlmodel